Add base colors to public API and extract to separate file
